requirements.txt
google-generativeai>=0.4.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.2.0
supabase>=2.3.0
asyncio>=3.4.3
aiofiles>=23.2.1
pytest>=7.0.0
pytest-cov>=4.0.0
tenacity>=8.2.0
